seperate stats into seperate file
